# perf: improve sql parsing performance