@davecramer 2015-06-03
Nice catch! the ?? thing has been a fiasco and ? never should have been allowed in the backend as an operator, but it is what it is:(