@vlsi 2015-06-03
Can we offload query parsing to the backend?

It looks currently we have to double parse the query (at jdbc and at backend) site.
Do you know if we can have "submit raw SQL" kind of feature in backend protocol?

I do understand it is not going to be solved in a matter of days.
Howeevr we can't implement full parsing logic at the JDBC side including custom operators, can we?