@davecramer 2015-06-03
Unfortunately not

There is no "raw sql" option.

I'm not going anywhere near custom operators. The ? was a poor decision.

Dave Cramer

On 3 June 2015 at 10:11, Vladimir Sitnikov <notifications@github.com> wrote:

> Can we offload query parsing to the backend?
>
> It looks currently we have to double parse the query (at jdbc and at
> backend) site.
> Do you know if we can have "submit raw SQL" kind of feature in backend
> protocol?
>
> I do understand it is not going to be solved in a matter of days.
> Howeevr we can't implement full parsing logic at the JDBC side including
> custom operators, can we?
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/pgjdbc/pgjdbc/pull/301#issuecomment-108443532>.
>