@vlsi 2015-06-03
> The ? was a poor decision.

I see.

The question is if we would have any luck in requesting such "raw SQL" API from backend developers.

At the end, what is their vision on how "driver developers" should handle "extendable operators". Do they understand "frontend" developers do not always have a choice which placeholder to use for the bind variable substitution?