@davecramer 2015-06-03
Well unless you are going to write it or fund it asking for features will
probably get you some discussion but not much else; that being said if you
don't ask the answer will certainly be no

How would you see raw SQL working ?

as for extendable operators, yes they are now aware of the issue; however
the decision was made quite a while ago. Feel free to enter the fray
http://www.postgresql.org/message-id/CA+Tgmoar8QHB2k15mZWvqa3LL81DgPDE_WiiqKLKnBfKQ71kDA@mail.gmail.com



Dave Cramer

On 3 June 2015 at 11:00, Vladimir Sitnikov <notifications@github.com> wrote:

> The ? was a poor decision.
>
> I see.
>
> The question is if we would have any luck in requesting such "raw SQL" API
> from backend developers.
>
> At the end, what is their vision on how "driver developers" should handle
> "extendable operators". Do they understand "frontend" developers do not
> always have a choice which placeholder to use for the bind variable
> substitution?
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/pgjdbc/pgjdbc/pull/301#issuecomment-108471383>.
>