@davecramer 2015-06-04
BTW, if you want to propose some changes to the backend regarding this there is an unconference in Ottawa. I can take the discussion to the developers there if you provide me with a coherent argument