@sehrope 2015-06-04
I really like the new interest in helping out with the driver but can we please stop including gigantic perf output in commit messages? It's making `git log` unusable.