@davecramer 2015-06-04
I'm going to have to agree here.

Dave Cramer

On 4 June 2015 at 18:06, Sehrope Sarkuni <notifications@github.com> wrote:

> I really like the new interest in helping out with the driver but can we
> please stop including gigantic perf output in commit messages? It's making git
> log unusable.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/pgjdbc/pgjdbc/pull/301#issuecomment-109069471>.
>