@vlsi 2015-06-04
Have you read this story on git commit messages? http://zachholman.com/posts/git-commit-history/