@vlsi 2015-06-04
By the way: is "Commit message format" https://github.com/pgjdbc/pgjdbc#commit-message-format still valid?

I adhere to that format and now you tell me that it is incorrect. How can it be the case?

Please update the guideline, so we all can save our time.