@davecramer 2015-06-04
I think the objection to the format is just the perf measurements?

Dave Cramer

On 4 June 2015 at 18:28, Vladimir Sitnikov <notifications@github.com> wrote:

> By the way: is "Commit message format"
> https://github.com/pgjdbc/pgjdbc#commit-message-format still valid?
>
> I adhere to that format and now you tell me that it is incorrect. How can
> it be the case?
>
> Please update the guideline, so we all can save our time.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/pgjdbc/pgjdbc/pull/301#issuecomment-109072766>.
>