@polobo 2015-06-04
On Thu, Jun 4, 2015 at 6:28 PM, Vladimir Sitnikov <notifications@github.com>
wrote:

> By the way: is "Commit message format"
> https://github.com/pgjdbc/pgjdbc#commit-message-format still valid?
>
> I adhere to that format and now you tell me that it is incorrect. How can
> it be the case?
>
​Structural adherence is one thing (and is correct per the guideline) but
the choice of what to include in the body is being questioned and, since
you bring it up, your choice to call this a "feature" instead of a
"performance" change(though the PR category is correct...)

> Please update the guideline, so we all can save our time.
>
Are you really blaming the guideline for not predicting the fact that
someone might decide to place perf output in the commit message and
proactively forbid it?​​

You can just as easily include the conclusion that you draw from the perf
run as a paragraph in the commit message without putting all of the raw
data there.  The raw data belongs either in the PR or, as I've suggested,
in a perf-log directory.  The guideline should indeed be updated once the
discussion has come to conclusion regarding how to handle this novel commit
message.

David J.