@vlsi 2015-06-04
> your choice to call this a "feature" instead of a "performance"

>You can just as easily include the conclusion that you draw from the perf
run as a paragraph in the commit message without putting all of the raw
data there

What will be the use of those conclusion?
Who could validate my conclusion?

I mean the following: there are lots and lots of cases when developers claim to "improve the performance", while in fact either their test bed was broken, or the conclusion was not correct.

You know, I am not anywhere near to the fine details. I think the subset of data I've included is the minimal possible subset that allows to cross-check my conclusions.

http://shipilev.net/blog/2015/black-magic-method-dispatch/